Revert "LPS-99657 Class that implements ConfigurationModelListener should be in package *.configuration.persistence.listener"
This reverts commit ec75bfa6eace8acf5368c5c8e5ca5aed9c03e494.
