+++ modules/apps/portal/portal-upload/src/main/java/com/liferay/portal/upload/internal/configuration/UploadServletRequestConfigurationModelListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.upload.internal.configuration.persistence.listener;
+package com.liferay.portal.upload.internal.configuration;
 
 import com.liferay.portal.configuration.persistence.listener.ConfigurationModelListener;
 import com.liferay.portal.configuration.persistence.listener.ConfigurationModelListenerException;
